requirements.txt
customtkinter==5.2.2
pillow==12.0.0
pdfplumber==0.11.8
python-docx==1.2.0
tkinterdnd2==0.4.3
pytest==9.0.2
